+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import EmotionRegistry from '@/components/provider/Registry';
 import React from 'react';
 import Header from '@/components/header/Header';
 import CssBaseline from '@mui/material/CssBaseline';
+import Background from '@/components/layout/Background';
 
 export const metadata: Metadata = {
   title: 'Cuculus Roadmap',
@@ -21,10 +22,12 @@ export default function RootLayout({
     <html lang="ja">
       <head />
       <CssBaseline />
-      <body style={{ backgroundColor: '#dcecfc' }}>
+      <body>
         <EmotionRegistry>
-          <Header />
-          {children}
+          <Background>
+            <Header />
+            {children}
+          </Background>
         </EmotionRegistry>
       </body>
     </html>
