+++ src/app/page.tsx
@@ -1,14 +1,10 @@
 import Caption from '@/app/_components/Caption';
-import Background from '@/app/_components/Background';
 
 export default function Home() {
   return (
     <main>
-      <Background>
-        <Caption />
-        <Caption /> {/*画像サイズ確認用*/}
-        <Caption /> {/*画像サイズ確認用*/}
-      </Background>
+      <Caption />
+      <div style={{ height: '1000px' }}>DUMMY</div>
     </main>
   );
 }
